Mostly submitted patches...
- Fix bug 54264 from Jon Smirl <jonsmirl@mediaone.net>
  Do cleanup of thread local storage on main thread.

- Fix bug 54275 from Jon Smirl <jonsmirl@mediaone.net>
  Release components in shell before shutting down xpcom

- Fix bug 54310 from Jon Smirl <jonsmirl@mediaone.net>
  Call JS_DestroyScript in xpcshell and js.c

- Fix bug 54352 from Jon Smirl <jonsmirl@mediaone.net>
  Cleanup what static data we can in xpclog.

- Initial fix of bug 54473
  Don't report warnings as errors in wrapped JS calls.

- Fix bug 54462 from Mark Hammond <MarkH@ActiveState.com>
  Fix jband's stupid use of uint8 for method indexes.

- Use environment rather than prefs for #ifdef'd debug options

- Don't report NS_ERROR_FACTORY_REGISTER_AGAIN as an error.

r=mccabe@netscape.com a=jband@netscape.com

+++ js/src/xpconnect/src/xpcprivate.h
@@ -72,6 +72,7 @@
 #include "xpcjsid.h"
 #include "prlong.h"
 #include "prmem.h"
+#include "prenv.h"
 
 #include "nsIJSContextStack.h"
 #include "prthread.h"
@@ -93,7 +94,6 @@
 
 #ifdef XPC_TOOLS_SUPPORT
 #include "nsIXPCToolsProfiler.h"
-#include "nsIPref.h"
 #endif
 
 #ifdef DEBUG
@@ -981,7 +981,7 @@ private:
                                  const nsXPTMethodInfo* method,
                                  const XPCNativeMemberDescriptor* desc,
                                  const nsXPTParamInfo& param,
-                                 uint8 vtblIndex,
+                                 uint16 vtblIndex,
                                  uint8 paramIndex,
                                  SizeMode mode,
                                  nsXPTCVariant* dispatchParams,
@@ -991,7 +991,7 @@ private:
                                      const nsXPTMethodInfo* method,
                                      const XPCNativeMemberDescriptor* desc,
                                      const nsXPTParamInfo& param,
-                                     uint8 vtblIndex,
+                                     uint16 vtblIndex,
                                      uint8 paramIndex,
                                      const nsXPTType& datum_type,
                                      nsXPTCVariant* dispatchParams,
